pull modal out of SiteRoleActions
