+++ src/core/client/admin/components/UserRole/SiteRoleActions.css
@@ -3,8 +3,3 @@
   justify-content: space-between;
   padding: 0;
 }
-
-.modal {
-  max-width: 500px;
-  padding: var(--spacing-2) var(--spacing-3) var(--spacing-3) var(--spacing-3);
-}
