+++ src/core/client/admin/components/UserRole/SiteRoleActionsModal.css
@@ -1,0 +1,4 @@
+.modal {
+  max-width: 500px;
+  padding: var(--spacing-2) var(--spacing-3) var(--spacing-3) var(--spacing-3);
+}
